# memory/ag_export/NIVXRAY_COMPLETE_AG_EXPORT/01_COMPLETE_SOURCE/backend/baselines/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyze": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/decode/chain": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
    "/api/decode/chain/export": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
    "/api/decode/chain/narrative": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
    "/api/decode/chain/narrative/stream": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
    "/api/decode/chain/split": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
    "/api/v2/analyze": {
      "post": {
        "has_request_body": true,
        "responses": [
          "200",
          "422"
        ]
      }
    },
...
  }
}